refactor: rename config.ts to runtime-config.ts and move static constants
- Rename `config.ts` to `runtime-config.ts` to clearly distinguish runtime configurations from build-time settings.
- Move environment-independent constants (`MISO_API_KEY`, `GCP_PROJECT_ID`, `GCP_LOG_NAME_PREFIX`) from the config file to `constant.ts`.
- Update all affected imports to reflect the file renaming and constant migration.
