+++ packages/mirror-tv-next/constants/runtime-config.ts
@@ -8,15 +8,9 @@
 const API_ENDPOINT_OVERRIDE_FROM_ENV = process.env.API_ENDPOINT
 const PROGRAMMABLE_SEARCH_API_KEY = process.env.PROGRAMMABLE_SEARCH_API_KEY
 const PROGRAMMABLE_SEARCH_ENGINE_ID = process.env.PROGRAMMABLE_SEARCH_ENGINE_ID
-const MISO_API_KEY = 'IHtn9b9tfPsO1EQpGV74OMf2syhELb6XVZe8u9FT'
-const GCP_PROJECT_ID = 'mirror-tv-275709'
-const GCP_LOG_NAME_PREFIX = 'mnews'
 
 export {
   API_ENDPOINT_OVERRIDE_FROM_ENV,
   PROGRAMMABLE_SEARCH_API_KEY,
   PROGRAMMABLE_SEARCH_ENGINE_ID,
-  MISO_API_KEY,
-  GCP_PROJECT_ID,
-  GCP_LOG_NAME_PREFIX,
 }
